chore: satisfy clippy -D warnings across the fabric
Type aliases for slot cells, entry-API handle inserts, inspect_err over
side-effecting map_err, collapsible conditions, redundant casts and
returns, and documented allowances where HANDLE-token APIs are
intentionally safe-wrapped. Compile-fail suite extended to three proofs:
native move-only, SharedRegion no-Clone, read-only mapping exposes no
mutable slice.

+++ experiments/authority-fabric/src/shared/windows.rs
@@ -63,7 +63,7 @@ pub fn duplicate_backing(file: &File, writable: bool) -> std::io::Result<File> {
             src,
             GetCurrentProcess(),
             &mut out,
-            access as u32,
+            access,
             0,
             0,
         );
@@ -128,7 +128,7 @@ pub fn map_read_write(file: &File, len: usize) -> std::io::Result<MappedReadWrit
     unsafe {
         let p = MapViewOfFile(
             file.as_raw_handle() as *mut winapi::ctypes::c_void,
-            FILE_MAP_WRITE as u32,
+            FILE_MAP_WRITE,
             0,
             0,
             0,
@@ -147,7 +147,7 @@ pub fn map_read_only(file: &File, len: usize) -> std::io::Result<MappedReadOnly>
     unsafe {
         let p = MapViewOfFile(
             file.as_raw_handle() as *mut winapi::ctypes::c_void,
-            FILE_MAP_READ as u32,
+            FILE_MAP_READ,
             0,
             0,
             0,
